backend/requirements.txt
# -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
# UniFi Backup Manager - Backend Requirements
# -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=

# FastAPI and ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Encryption
cryptography>=42.0.0

# HTTP Client (for UniFi API)
aiohttp>=3.9.0

# Scheduling
apscheduler>=3.10.0

# Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.1.0

# Date handling
python-dateutil>=2.8.0
